Update api review book

+++ app/Repositories/BookRepositoryEloquent.php
@@ -273,22 +273,26 @@ class BookRepositoryEloquent extends AbstractRepositoryEloquent implements BookR
     public function review($bookId, array $data)
     {
         $book = $this->model()->findOrFail($bookId);
-        $dataReview = array_only($data, ['content', 'star']);
+        $dataReview = array_only($data, ['content', 'star', 'owner_id']);
         $dataReview['created_at'] = $dataReview['updated_at'] = Carbon::now();
 
-        $book->reviews()->detach($this->user->id);
-
-        $book->reviews()->attach([
-            $this->user->id => $dataReview
-        ]);
-
-        if (isset($dataReview['star'])) {
-            Event::fire('books.averageStar', [
-                [
-                    'book' => $book,
-                    'star' => $dataReview['star'],
-                ]
+        if (in_array($dataReview['owner_id'], $book->owners->pluck('id')->toArray())) {
+            $book->reviews()->detach($this->user->id);
+            $book->reviews()->attach([
+                $this->user->id => $dataReview
             ]);
+
+            if (isset($dataReview['star'])) {
+                Event::fire('books.averageStar', [
+                    [
+                        'book' => $book,
+                        'owner_id' => $dataReview['owner_id'],
+                        'star' => $dataReview['star'],
+                    ]
+                ]);
+            }
+        } else {
+            throw new ActionException(__FUNCTION__);
         }
     }
